delete early version of useJobEvents, rename files accordingly
